Website Update
Completed steps 1 - 18 by displaying the body as a flexbox, setting the growth, shrink, and basis for specific elements, and set the order of the footer so it it below the rest of the elements. Still need to complete the last few steps and make sure all links and buttons function.